+++ src/util.h
@@ -13,7 +13,7 @@
 #include "config.h"
 
 
-#define ARR_LEN(array) (sizeof(array) / sizeof(array[0]))
+#define ARR_LEN(array) (int)(sizeof(array) / sizeof(array[0]))
 
 
 char* ml_util_buf2str(const char *buf, size_t len);
